// src/api/Shop/endpoints.ts
export enum ShopApi {
  GetShopsNames = 'Shop/GetShopNames',
  GetAllBranches = 'Shop/GetAllBranches',
  GetBranchById = 'Shop/GetBranchById',
  GetBranchNames = 'Shop/GetBranchNames',
  AddShopWithBranch = 'Shop/AddShopWithBranch',
  AddBranch = 'Shop/AddBranch',
  ModifyShop = 'Shop/ModifyShop',
  ModifyBranch = 'Shop/ModifyBranch',
  DeleteBranch = 'Shop/DeleteBranch',
  GetBranchNamesWithProduct = 'Shop/GetBranchNamesWithProduct',
}